requirements.txt
coverage==4.5.2
Django==2.1.3
django-autofixture==0.12.1
django-jenkins==0.110.0
django-modern-rpc==0.11.1
django-seed==0.1.9
djangorestframework==3.9.0
Faker==1.0.0
mysqlclient==1.3.13
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyotp==2.2.7
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
six==1.11.0
text-unidecode==1.2
